# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Focusly Backend API",
    "description": "Local AI Learning Platform converting documents into summaries, quizzes, and narrated videos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Users Admin",
        "description": "Lists all platform users along with their monthly usage statistics.\nRequires user.is_admin flag.",
        "operationId": "list_all_users_admin_api_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/admin/usage": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get User Usage Stats",
        "description": "Returns aggregate usage statistics across all platform users.",
        "operationId": "get_user_usage_stats_api_admin_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/admin/logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get System Logs Admin",
        "description": "Returns recent system logs, connection events, and error records.",
        "operationId": "get_system_logs_admin_api_admin_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "INFO",
              "title": "Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/maintenance/cleanup": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Run Maintenance Cleanup",
        "description": "Cleans expired share links and temporary upload files in /tmp/focusly_uploads.",
        "operationId": "run_maintenance_cleanup_api_admin_maintenance_cleanup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/user/usage": {
      "get": {
        "summary": "Get User Usage Endpoint",
        "description": "Returns current month usage counters, monthly limits, and next reset date.",
        "operationId": "get_user_usage_endpoint_api_user_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUsageStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/exam-mode/start": {
      "post": {
        "summary": "Start Timed Exam Endpoint",
        "description": "Combines documents and generates a long-form timed test (30-50 questions, ~2 min/question time limit).",
        "operationId": "start_timed_exam_endpoint_api_exam_mode_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimedExamStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exam-mode/status": {
      "get": {
        "summary": "Get Timed Exam Status Endpoint",
        "description": "Returns time_remaining, current_question, total_questions, and progress.",
        "operationId": "get_timed_exam_status_endpoint_api_exam_mode_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exam-mode/pause": {
      "post": {
        "summary": "Pause Timed Exam Endpoint",
        "description": "Pauses active exam timer (resumable within 24 hours).",
        "operationId": "pause_timed_exam_endpoint_api_exam_mode_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exam-mode/resume": {
      "post": {
        "summary": "Resume Timed Exam Endpoint",
        "description": "Resumes paused exam timer.",
        "operationId": "resume_timed_exam_endpoint_api_exam_mode_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/exam-mode/submit": {
      "post": {
        "summary": "Submit Timed Exam Endpoint",
        "description": "Grades timed exam test, updates weak areas, and returns score and feedback.",
        "operationId": "submit_timed_exam_endpoint_api_exam_mode_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimedExamSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/study/ask": {
      "post": {
        "summary": "Ask Focusly Endpoint",
        "description": "Ask Focusly AI Q&A chatbot grounded in student study documents and learning history.",
        "operationId": "ask_focusly_endpoint_api_study_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/documents/upload-image": {
      "post": {
        "summary": "Upload Image Notes Endpoint",
        "description": "Uploads 1 to 10 handwritten note images (JPEG, PNG, HEIC, WebP).\nUploads images to Supabase Storage, creates document record with status='processing',\nand triggers background Celery task `process_image_notes_task`.",
        "operationId": "upload_image_notes_endpoint_api_documents_upload_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_notes_endpoint_api_documents_upload_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageNotesUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/{document_id}/status": {
      "get": {
        "summary": "Get Document Status Endpoint",
        "description": "Returns detailed document status including OCR confidence, virtual page map, progress, and error message.",
        "operationId": "get_document_status_endpoint_api_documents__document_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/upload": {
      "post": {
        "summary": "Upload Document",
        "description": "Validates PDF file, uploads to Supabase Storage, extracts text via PyMuPDF/OCR/pdfplumber,\nstores record in database, and sets status='completed'.",
        "operationId": "upload_document_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Document Documents Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_documents_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserUsageStatsResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "month": {
            "type": "integer",
            "title": "Month"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "video_generations_used": {
            "type": "integer",
            "title": "Video Generations Used"
          },
          "video_generations_limit": {
            "type": "integer",
            "title": "Video Generations Limit",
            "default": 10
          },
          "summary_generations_used": {
            "type": "integer",
            "title": "Summary Generations Used"
          },
          "summary_generations_limit": {
            "type": "integer",
            "title": "Summary Generations Limit",
            "default": 50
          },
          "quiz_generations_used": {
            "type": "integer",
            "title": "Quiz Generations Used",
            "default": 0
          },
          "quiz_generations_limit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiz Generations Limit",
            "default": "unlimited"
          },
          "next_reset_date": {
            "type": "string",
            "title": "Next Reset Date"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "This content was generated by AI locally. Usage limits apply monthly."
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "month",
          "year",
          "video_generations_used",
          "summary_generations_used",
          "next_reset_date"
        ],
        "title": "UserUsageStatsResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}